Add deep link support for Android

+++ Auth/src/commonMain/kotlin/io/github/jan/supacompose/auth/user/UserSession.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class UserSession(
     @SerialName("access_token") val accessToken: String,
     @SerialName("refresh_token") val refreshToken: String,
-    @SerialName("expires_in") val expiresIn: Int,
+    @SerialName("expires_in") val expiresIn: Long,
     @SerialName("token_type") val tokenType: String,
     val user: UserInfo?,
     @SerialName("type") val type: String = "",
